+++ Runtime/UI/Components/Accordion.cs
@@ -9,6 +9,33 @@ using Unity.Properties;
 namespace Unity.AppUI.UI
 {
     /// <summary>
+    /// The position of an element inside a Flex container.
+    /// </summary>
+    [GenerateLowerCaseStrings]
+    public enum FlexPosition
+    {
+        /// <summary>
+        /// The element is at the start of the container.
+        /// </summary>
+        /// <remarks>
+        /// In a row with Left-to-Right layout, this is the left side.
+        /// In a row with Right-to-Left layout, this is the right side.
+        /// For a column, this is the top side.
+        /// </remarks>
+        Start,
+        
+        /// <summary>
+        /// The element is at the end of the container.
+        /// </summary>
+        /// <remarks>
+        /// In a row with Left-to-Right layout, this is the right side.
+        /// In a row with Right-to-Left layout, this is the left side.
+        /// For a column, this is the bottom side.
+        /// </remarks>
+        End
+    }
+    
+    /// <summary>
     /// Item used inside an <see cref="Accordion"/> element.
     /// </summary>
 #if ENABLE_UXML_SERIALIZED_DATA
@@ -18,6 +45,8 @@ namespace Unity.AppUI.UI
     {
 #if ENABLE_RUNTIME_DATA_BINDINGS
         internal static readonly BindingId titleProperty = nameof(title);
+
+        internal static readonly BindingId indicatorPositionProperty = nameof(indicatorPosition);
         
         internal static readonly BindingId valueProperty = nameof(value);
         
@@ -65,12 +94,20 @@ namespace Unity.AppUI.UI
         /// The AccordionItem heading styling class.
         /// </summary>
         public const string headingUssClassName = ussClassName + "__heading";
+        
+        /// <summary>
+        /// The AccordionItem indicator position styling class.
+        /// </summary>
+        [EnumName("GetIndicatorPosUssClassName", typeof(FlexPosition))]
+        public const string indicatorPosUssClassName = ussClassName + "--indicator-";
 
         readonly VisualElement m_ContentElement;
         
         readonly VisualElement m_ContentParentElement;
 
         readonly LocalizedTextElement m_HeaderTextElement;
+
+        readonly VisualElement m_HeaderIndicatorElement;
 
         readonly Pressable m_Clickable;
 
@@ -94,8 +131,8 @@ namespace Unity.AppUI.UI
             trailingContainer = new VisualElement { name = trailingContainerUssClassName, pickingMode = PickingMode.Ignore };
             trailingContainer.AddToClassList(trailingContainerUssClassName);
 
-            var headerIndicatorElement = new Icon { name = indicatorUssClassName, iconName = k_IndicatorIconName, pickingMode = PickingMode.Ignore };
-            headerIndicatorElement.AddToClassList(indicatorUssClassName);
+            m_HeaderIndicatorElement = new Icon { name = indicatorUssClassName, iconName = k_IndicatorIconName, pickingMode = PickingMode.Ignore };
+            m_HeaderIndicatorElement.AddToClassList(indicatorUssClassName);
 
             m_HeaderElement = new ExVisualElement
             {
@@ -110,7 +147,7 @@ namespace Unity.AppUI.UI
             m_HeaderElement.AddManipulator(new KeyboardFocusController(OnKeyboardFocus, OnFocus));
             m_HeaderElement.hierarchy.Add(m_HeaderTextElement);
             m_HeaderElement.hierarchy.Add(trailingContainer);
-            m_HeaderElement.hierarchy.Add(headerIndicatorElement);
+            m_HeaderElement.hierarchy.Add(m_HeaderIndicatorElement);
 
             var headingElement = new VisualElement { pickingMode = PickingMode.Ignore };
             headingElement.AddToClassList(headingUssClassName);
@@ -135,6 +172,7 @@ namespace Unity.AppUI.UI
             hierarchy.Add(headingElement);
             hierarchy.Add(m_ContentParentElement);
             
+            AddToClassList(GetIndicatorPosUssClassName(FlexPosition.End));
             SetValueWithoutNotify(false);
         }
         
@@ -191,6 +229,39 @@ namespace Unity.AppUI.UI
 #if ENABLE_RUNTIME_DATA_BINDINGS
                 if (changed)
                     NotifyPropertyChanged(in trailingContentTemplateProperty);
+#endif
+            }
+        }
+        
+        /// <summary>
+        /// The position of the indicator.
+        /// </summary>
+        [Tooltip("The position of the indicator.")]
+#if ENABLE_RUNTIME_DATA_BINDINGS
+        [CreateProperty]
+#endif
+#if ENABLE_UXML_SERIALIZED_DATA
+        [UxmlAttribute]
+#endif
+        public FlexPosition indicatorPosition
+        {
+            get => m_HeaderElement.hierarchy.IndexOf(m_HeaderIndicatorElement) == 0 ? FlexPosition.Start : FlexPosition.End;
+            set
+            {
+                var previousValue = indicatorPosition;
+                if (previousValue == value)
+                    return;
+                
+                m_HeaderIndicatorElement.RemoveFromHierarchy();
+                if (value == FlexPosition.Start)
+                    m_HeaderElement.hierarchy.Insert(0, m_HeaderIndicatorElement);
+                else
+                    m_HeaderElement.hierarchy.Add(m_HeaderIndicatorElement);
+                RemoveFromClassList(GetIndicatorPosUssClassName(previousValue));
+                AddToClassList(GetIndicatorPosUssClassName(value));
+                
+#if ENABLE_RUNTIME_DATA_BINDINGS
+                NotifyPropertyChanged(indicatorPositionProperty);
 #endif
             }
         }
@@ -292,6 +363,12 @@ namespace Unity.AppUI.UI
                 defaultValue = "Header",
             };
             
+            readonly UxmlEnumAttributeDescription<FlexPosition> m_IndicatorPosition = new UxmlEnumAttributeDescription<FlexPosition>
+            {
+                name = "indicator-position",
+                defaultValue = FlexPosition.End,
+            };
+            
             readonly UxmlBoolAttributeDescription m_Value = new UxmlBoolAttributeDescription
             {
                 name = "value",
@@ -311,6 +388,7 @@ namespace Unity.AppUI.UI
 
                 var element = (AccordionItem)ve;
                 element.title = m_Title.GetValueFromBag(bag, cc);
+                element.indicatorPosition = m_IndicatorPosition.GetValueFromBag(bag, cc);
                 element.value = m_Value.GetValueFromBag(bag, cc);
             }
         }