+++ Runtime/UI/Manipulators/Draggable.cs
@@ -108,6 +108,11 @@ namespace Unity.AppUI.UI
             if (active)
                 target?.Blur();
         }
+        
+        /// <summary>
+        /// Accept the drag operation.
+        /// </summary>
+        public Func<Draggable, bool> acceptDrag { get; set; }
 
         /// <summary>
         /// This method processes the down event sent to the target Element.
@@ -117,6 +122,10 @@ namespace Unity.AppUI.UI
         /// <param name="pointerId"> The pointer id of the pointer.</param>
         protected override void ProcessDownEvent(EventBase evt, Vector2 localPosition, int pointerId)
         {
+            var canDrag = acceptDrag?.Invoke(this) ?? true;
+            if (!canDrag)
+                return;
+            
             deltaPos = Vector2.zero;
             deltaStartPos = Vector2.zero;
             this.localPosition = localPosition;