+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-pre.5] - 2024-06-18
+
+### Changed
+
+- Complete rewrite of the SplitView component. The SplitView is no more a derived from TwoPaneSplitView from UI-Toolkit, but a full custom component that supports any number of panes.
+
+### Fixed
+
+- Fixed reset of Dropdown value when changing its source items.
+- Fixed a visual bug where the checkmark symbol didn't appear on DropdownItem or MenuItem that have a `selected` state.
+
+### Added
+
+- Added `indicatorPosition` property in `AccordionItem` component, in order to swap the indicator position either at start or end of the heading row.
+- Added `check` regular icon as a required icon in App UI themes.
+
 ## [2.0.0-pre.4] - 2024-06-05
 
 ### Fixed